+++ alicloud/provider.go
@@ -68,6 +68,7 @@ func Provider() terraform.ResourceProvider {
 			"alicloud_security_group_rule":       resourceAliyunSecurityGroupRule(),
 			"alicloud_db_account":                resourceAlicloudDBAccount(),
 			"alicloud_db_account_privilege":      resourceAlicloudDBAccountPrivilege(),
+			"alicloud_db_connection":             resourceAlicloudDBConnection(),
 			"alicloud_db_instance":               resourceAlicloudDBInstance(),
 			"alicloud_ess_scaling_group":         resourceAlicloudEssScalingGroup(),
 			"alicloud_ess_scaling_configuration": resourceAlicloudEssScalingConfiguration(),